adding transport Microservice

+++ services/clientCRS.service.js
@@ -3,28 +3,40 @@ const EXCHANGE_NAME = "Client";
 const orderRepository = require('../repository/order.repository');
 const logRepository = require('../repository/logClient.repository');
 
+exports.new = async (body, res) => {
+    const order =  await orderRepository.new(body);
 
-exports.new = async (body) => {
-    const newLog = await logRepository.newCRS("Order created.");
+    const newLog = await logRepository.newCRS(`Order created :${order._id}.`);
 
-    await orderRepository.new(body);
+    rabbitSender.send(EXCHANGE_NAME, newLog.log);
 
-    rabbitSender.send(EXCHANGE_NAME, newLog.log  );
+    res.status(201).send(newLog.log);
 };
 
-exports.cancel = async (id) => {
-    const newLog = await logRepository.newCRS("Order deleted.");
+exports.cancel = async (id, res) => {
+    const order = await orderRepository.get(id);
+    if(order.status == "updated" || order.status == "created") {
+        const newLog = await logRepository.newCRS(`Order deleted :${id}.`);
 
-    await orderRepository.delete(id);
+        await orderRepository.delete(id);
 
-    rabbitSender.send(EXCHANGE_NAME, newLog.log  );
+        rabbitSender.send(EXCHANGE_NAME, newLog.log  );
+        res.status(202).send(newLog.log);
+    } else {
+        res.status(400).send(`Cant cancel an order when its ${order.status}`);
+    }
 };
 
-exports.update = async (id, body) => {
-    const newLog = await logRepository.newCRS("Order updated.");
+exports.update = async (id, body, res) => {
+    const order = await orderRepository.get(id);
+    if(order.status == "updated" || order.status == "created") {
+        const newLog = await logRepository.newCRS(`Order updated :${id}.`);
 
-    await orderRepository.update(id, body);
+        await orderRepository.update(id, body);
 
-    rabbitSender.send(EXCHANGE_NAME, newLog.log  );
+        rabbitSender.send(EXCHANGE_NAME, newLog.log  );
+        res.status(202).send(newLog.log);
+    } else {
+        res.status(400).send(`Cant update an order when its ${order.status}`);
+    }
 };
-
